feat: avoid using enums in files to build docs
