+++ tests/endpoints/test_voyages_search_enriched.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 from tests.testcases import TestCaseUsingRealAPI
 from vortexasdk import VoyagesSearchEnriched
-from vortexasdk.api.voyages import VoyageDateRangeActivity
+from vortexasdk.api.shared_types import VoyageDateRangeActivity
 
 rotterdam = "68faf65af1345067f11dc6723b8da32f00e304a6f33c000118fccd81947deb4e"
