+++ vortexasdk/api/voyages.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from pydantic import BaseModel
 from typing import List, Optional, Union
 from typing_extensions import Literal
@@ -195,19 +194,3 @@ class VoyageEnrichedItem(BaseModel):
     ] = None
     latest_product_details: Optional[List[List[EntityWithSingleLayer]]] = None
     voyage_status: str
-
-
-class VoyageDateRangeActivity(Enum):
-    ACTIVE = "active"
-    DEPARTURES = "departures"
-    ARRIVALS = "arrivals"
-
-
-class OriginBehaviour(Enum):
-    ANY_LOAD = "any_load"
-    FIRST_LOAD = "first_load"
-
-
-class DestinationBehaviour(Enum):
-    ANY_LOAD = "any_discharge"
-    LAST_DISCHARGE = "last_discharge"
