+++ vortexasdk/endpoints/voyages_congestion_breakdown.py
@@ -7,18 +7,13 @@ from datetime import datetime
 from typing import Any, Dict, List, Union
 
 from vortexasdk.api import ID
-from vortexasdk.api.shared_types import Tag, to_ISODate
+from vortexasdk.api.shared_types import Tag, to_ISODate, VoyageDateRangeActivity, OriginBehaviour, DestinationBehaviour
 from vortexasdk.endpoints.endpoints import VOYAGES_CONGESTION_BREAKDOWN
 from vortexasdk.endpoints.voyages_congestion_breakdown_result import (
     CongestionBreakdownResult,
 )
 from vortexasdk.operations import Search
 from vortexasdk.utils import convert_to_list
-from vortexasdk.api.voyages import (
-    VoyageDateRangeActivity,
-    OriginBehaviour,
-    DestinationBehaviour,
-)
 
 
 class VoyagesCongestionBreakdown(Search):
