+++ vortexasdk/endpoints/voyages_product_breakdown.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from typing import Any, Dict, List, Union
 
 from vortexasdk.api import ID
-from vortexasdk.api.shared_types import Tag, to_ISODate
+from vortexasdk.api.shared_types import Tag, to_ISODate, VoyageDateRangeActivity, OriginBehaviour, DestinationBehaviour
 
 from vortexasdk.endpoints.endpoints import VOYAGES_BREAKDOWN
 from vortexasdk.endpoints.voyages_breakdown_result import (
@@ -16,11 +16,6 @@ from vortexasdk.endpoints.voyages_breakdown_result import (
 
 from vortexasdk.operations import Search
 from vortexasdk.utils import convert_to_list
-from vortexasdk.api.voyages import (
-    VoyageDateRangeActivity,
-    OriginBehaviour,
-    DestinationBehaviour,
-)
 
 # noinspection PyUnresolvedReferences
