+++ vortexasdk/endpoints/voyages_top_hits.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from typing import Any, Dict, List, Union
 
 from vortexasdk.api import ID
-from vortexasdk.api.shared_types import Tag, to_ISODate
+from vortexasdk.api.shared_types import Tag, to_ISODate, VoyageDateRangeActivity, OriginBehaviour, DestinationBehaviour
 from vortexasdk.endpoints.aggregation_breakdown_result import (
     AggregationBreakdownResult,
 )
@@ -15,11 +15,6 @@ from vortexasdk.endpoints.endpoints import VOYAGES_TOP_HITS
 
 from vortexasdk.operations import Search
 from vortexasdk.utils import convert_to_list
-from vortexasdk.api.voyages import (
-    VoyageDateRangeActivity,
-    OriginBehaviour,
-    DestinationBehaviour,
-)
 
 
 class VoyagesTopHits(Search):
